app: Move actions service into workflow module
