+++ tracecat/types/api.py
@@ -2,62 +2,9 @@ from __future__ import annotations
 
 from typing import Any, Literal
 
-from pydantic import BaseModel, Field
+from pydantic import BaseModel
 
 from tracecat.db.schemas import Resource
-from tracecat.dsl.enums import JoinStrategy
-from tracecat.dsl.models import ActionRetryPolicy
-
-# TODO: Consistent API design
-# Action and Workflow create / update params
-# should be the same as the metadata responses
-
-RunStatus = Literal["pending", "running", "failure", "success", "canceled"]
-
-
-class ActionControlFlow(BaseModel):
-    run_if: str | None = None
-    for_each: str | list[str] | None = None
-    retry_policy: ActionRetryPolicy = Field(default_factory=ActionRetryPolicy)
-    start_delay: float = Field(
-        default=0.0, description="Delay before starting the action in seconds."
-    )
-    join_strategy: JoinStrategy = Field(default=JoinStrategy.ALL)
-
-
-class ActionResponse(BaseModel):
-    id: str
-    type: str
-    title: str
-    description: str
-    status: str
-    inputs: dict[str, Any]
-    key: str  # Computed field
-    control_flow: ActionControlFlow = Field(default_factory=ActionControlFlow)
-
-
-class ActionMetadataResponse(BaseModel):
-    id: str
-    workflow_id: str
-    type: str
-    title: str
-    description: str
-    status: str
-    key: str
-
-
-class CreateActionParams(BaseModel):
-    workflow_id: str
-    type: str
-    title: str
-
-
-class UpdateActionParams(BaseModel):
-    title: str | None = None
-    description: str | None = None
-    status: str | None = None
-    inputs: dict[str, Any] | None = None
-    control_flow: ActionControlFlow | None = None
 
 
 class UpsertWebhookParams(BaseModel):
